add roles abiturients and news and forum

+++ app/controllers/sections_controller.rb
@@ -1,22 +1,25 @@
 class SectionsController < ApplicationController
      def create
-      @section = Section.new(params[:section])
-      if @section.save
-        redirect_to forum_path
-      else
-        render :action => :new
+      if current_user && current_user.role && current_user.role.can_admin_forum
+        @section = Section.new(params[:section])
+        if @section.save == false
+          render :action => :new
+        end
+      end
+      redirect_to forum_path
       end
 
 end
   def destroy
-    @Section = Section.find(params[:id])
-    @Section.theme.each do|n|
+    if current_user && current_user.role && current_user.role.can_admin_forum
+      @Section = Section.find(params[:id])
+      @Section.theme.each do|n|
       n.post.each do|t|
         t.destroy
       end
      n.destroy
     end
     @Section.destroy
-    redirect_to forum_path
   end
-      end
+    redirect_to forum_path
+ end
