+++ app/controllers/themes_controller.rb
@@ -1,19 +1,22 @@
 class ThemesController < ApplicationController
   def create
+    if current_user && current_user.role && current_user.role.can_admin_forum
       @theme = Theme.new(params[:theme])  
-     if @theme.save
-        redirect_to forum_path
-      else
+      if @theme.save == false
         render :controller=>"forum", :action => :new_theme
       end
+      redirect_to forum_path
     end
+  end
+end
     
-  def destroy
+def destroy
+  if current_user && current_user.role && current_user.role.can_admin_forum
     @Theme = Theme.find(params[:id])
     @Theme.post.each do|t|
-      t.destroy
+     t.destroy
     end
     @Theme.destroy
-    redirect_to forum_path
   end
+    redirect_to forum_path
 end
